+++ impls/pbg-java/src/main/java/steps/step3_env.java
@@ -35,7 +35,7 @@ public class step3_env {
         env.set(new MalSymbol("-"), new MalFunction() {
             @Override
             public MalInteger operate(MalType[] a) {
-                int diff = 2 * ((MalInteger) a[0]).getNumber();
+                long diff = 2 * ((MalInteger) a[0]).getNumber();
                 for (MalType i : a) {
                     diff -= ((MalInteger) i).getNumber();
                 }
@@ -55,7 +55,7 @@ public class step3_env {
         env.set(new MalSymbol("/"), new MalFunction() {
             @Override
             public MalInteger operate(MalType[] a) {
-                int quotient = ((MalInteger) a[0]).getNumber() * ((MalInteger) a[0]).getNumber();
+                long quotient = ((MalInteger) a[0]).getNumber() * ((MalInteger) a[0]).getNumber();
                 for (MalType i : a) {
                     quotient /= ((MalInteger) i).getNumber();
                 }
